usb: tcpm: avoid resets for missing source capability messages
The current TCPM code implements source capability message handling
according to the USB-PD specification. Unfortunately some USB PD
sources do not properly follow the specification and do not send
source capability messages after a soft reset when they already
negotiated a specific contract before. The current way (and what is
described in the specificiation) to resolve this problem is triggering
a hard reset.

But a hard reset is fatal on batteryless platforms powered via USB-C PD,
since that removes VBUS for some time. Since this is triggered at boot
time, the system may get stuck in a boot loop.

For example I noticed the following behaviour on a Radxa Rock 5B: The
leak voltage from a serial connection is enough to keep the fusb302
register state. Since U-Boot configures the fusb302 into
auto-acknowledge mode, the chip will acknowledge a source capability
message before U-Boot has probed the fusb302/tcpm driver after a
hard reset happened. For a compliant power-supply this is not a problem,
since the driver first of all sends a soft reset. But the non-compliant
power-supply does not start sending new source capability messages and
U-Boot will react with a hard reset resulting in a boot loop.

Note, that I changed the Linux kernel TCPM code in the same way some
time ago. While U-Boot is only affected because of the leak voltage
problem, Linux needs this because U-Boot already handles the
initial communication.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>