board: rockchip: add fusb302 node on Sige5
The Sige5 uses two fusb302 controllers, one of which is on the USB Type
C power input connector. Add this controller to the Sige5's u-boot
specific tree, as it is critical that it is initialised before Linux
takes over, so that it does not kill the power input once someone adds
the fusb302 node to its device tree.

Also enable the necessary FUSB302 u-boot driver in the Sige5's defconfig.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>