net: dwc: Increase Rockchip DMA SW reset timeout to 2 seconds
On warm reboot on the Rock 4d, the DMA controller takes a long time to
complete the SW reset, leading to messages like:

    EQOS_DMA_MODE_SWR stuck
    FAILED: -110
    Could not start ethernet@2a220000

On linux, some hardware that use this controller set the reset timeout
to up to 2 seconds (See f06e4bfd010faefa637689d2df2c727dbf6e1d27).

Still on linux, from what I see, in the case of rockchip, the reset
function of the DMA is never called, that's why it doesn't need any
special treatemnt in linux.

In u-boot, we could just skip the SW reset altogether but the educated
guess of using a 2 second timeout (instead of 50 ms) also seem to fix the
issue. (Documentation is currently limited)

To avoid adding code to skip the SW reset for specific hardware, which
could be wrong anyway, this commit just increases the SW reset timeout to
2 seconds for rockchip HW.

Signed-off-by: Detlev Casanova <detlev.casanova@collabora.com>