usb: tcpm: fix pd_transmit poll condition
The read_poll_timeout() exit "cond" parameter wait for a complete
transmit state.

The "!tx_complete" cond means that on the first tcpm_transmit_helper()
call, if the tx_complete returns to false, the poll loop would exit.
But since at probe the tx_status is left at 0, the tx is considered
as successful. And then later the proper status is returned and
ignored.

But if on the first tcpm_transmit_helper() call the TX status is
returned the read_poll_timeout() would return in timeout and the
negociation would fail.

Fixes: 1db4c0ac77e ("usb: tcpm: add core framework")
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>
Tested-by: Marcel Ziswiler <marcel.ziswiler@codethink.co.uk> # rock5b
Link: https://lore.kernel.org/r/20250526-topic-tcpm-tx-poll-cond-v1-1-38a81560ffff@linaro.org
Signed-off-by: Sebastian Reichel <sre@ring0.de>